Fix versioning
After adding the git tag "v1.0" the old method for obtaining
the git hash stopped working resulting in versions like
"v1.0v1.0". This has been fixed now. The version is now
the ouput of 'git describe --always --long --dirty' and
includes the last tag, the number of commits since that
tag, the git hash of the last commit, and a dirty flag
in the event of the build tree is spoilt.