frugen: Add building of static binary
Sometimes it is needed to use frugen on a different machine
than it was built on. Sometimes those machines don't have
the same versions of shared libraries.

For ease of binary portability this commit adds building
of `frugen-static` binary in addition to the usual `frugen`.
That resolves dependency on certain versions of glibc and libjson-c.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>