uuid: Fix byte order
According to IPMI FRU Information Storage Specification,
the System Unique ID field in the Multirecord area is a
"copy of the system GUID from [SMBIOS]", which means
it must follow the DMTF SMBIOS Specification.

The specification prescribes little endian encoding
for the first three parts of UUID.

This commit add that functionality and reorders the
bytes as presribed by the specs.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>