fru_get_typelen: Fix autodetection of binary
TAB, CR and LF symbols were not considered text
and were forcing binary encoding to be used.
That is fixed now. Any data containing pure text
and the above characters will now be reported as
8-bit ASCII (plain text).

frugen will not encode text with those characters
as binary as it would before this commit.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>